+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+* Validation check in `flexberry-edit-form` blueprint.
 
 ## [2.5.0-beta.7] - 2020-07-22
 ### Added